Add new syntax for nesting objects in INI config files
When a section name starts with a dot, it will be nested in the object
created by the previous top level section:

    [base]
    x = 1
    [.nested]
    y = 2

This will result in the following object:

    {"base": {"x": 1, "nested": {"y": 2}}}

Currently, only one level of nesting is supported. I'm considering adding
multiple levels of nesting using multiple dots, but this is a bit more
complex and also leads to duplicate section errors in `configparser`.
Duplicate sections can be allowed, but that opens another can of worms.

A think a solution might be to allow duplicate sections in the config
parser and then check to be sure keys aren't being overwritten in any
given object or subobject.